app/requirements.txt
python-dotenv
mysql-connector-python
firebase-admin
streamlit
